requirements.txt
flask>=3.0.0
flask-compress>=1.13
gunicorn>=21.0.0
playwright>=1.40.0
anthropic>=0.18.0
openai>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
lxml>=5.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
